migra. almacen

+++ database/migrations/2024_12_20_135404_create_copas.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('codigo_botella');
             $table->integer('codigo_copa');
-            $table->smallInteger('equivalencia');
+            $table->decimal('equivalencia',10,2);
             $table->timestamps();
         });
     }
